requirements.txt
streamlit
gspread
google-auth
openai
